+++ service/controller/resource/skip/delete.go
@@ -1,0 +1,17 @@
+package skip
+
+import (
+	"context"
+
+	"github.com/giantswarm/microerror"
+)
+
+func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	// Call r.EnsureCreated to avoid repetition as the logic is the same.
+	err := r.EnsureCreated(ctx, obj)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}