+++ service/controller/resource/skip/doc.go
@@ -1,0 +1,3 @@
+package skip
+
+// Package skip implements a resource which allow to skip the entire reconciliation for a specific object based on its name.